- Each user is supplied with their own entry list (entries associated to users) - entry functions extended to account for username (Principal user.getName()) - user details get stripped from entries before responding

+++ src/main/java/ch/zli/m223/punchclock/service/EntryService.java
@@ -1,7 +1,9 @@
 package ch.zli.m223.punchclock.service;
 
 import ch.zli.m223.punchclock.domain.Entry;
+import ch.zli.m223.punchclock.domain.PunchClockUser;
 import ch.zli.m223.punchclock.repository.EntryRepository;
+import ch.zli.m223.punchclock.repository.UserRepository;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -9,34 +11,40 @@ import java.util.List;
 @Service
 public class EntryService {
     private EntryRepository entryRepository;
+    private UserRepository userRepository;
 
-    public EntryService(EntryRepository entryRepository) {
+    public EntryService(EntryRepository entryRepository, UserRepository userRepository) {
         this.entryRepository = entryRepository;
+        this.userRepository = userRepository;
     }
 
-    public Entry createEntry(Entry entry) {
-        return entryRepository.saveAndFlush(entry);
+    public Entry createEntry(String username, Entry entry) {
+        entry.setUser(userRepository.findByUsername(username));
+        Entry savedEntry = entryRepository.saveAndFlush(entry);
+        savedEntry.setUser(null);
+        return savedEntry;
     }
 
-    public Boolean deleteEntry(long entryId) {
+    public Boolean deleteEntry(String username, long entryId) {
         if (entryRepository.existsById(entryId)) {
-            entryRepository.deleteById(entryId);
+            entryRepository.deleteDistinctByUser_UsernameAndId(username, entryId);
             return true;
         }
         return false;
     }
 
-    public Entry updateEntry(long entryId, Entry entry) {
-        if (entry.getId() == entryId) {
-            if (entryRepository.existsById(entryId)) {
-                entryRepository.deleteById(entryId);
-                return entryRepository.saveAndFlush(entry);
-            }
+    public Entry updateEntry(String username, long entryId, Entry entry) {
+        entry.setId(entryId);
+        if (entryRepository.existsById(entryId) && entry.getUser().getUsername().equals(username)) {
+            entryRepository.deleteById(entryId);
+            Entry updatedEntry = entryRepository.saveAndFlush(entry);
+            updatedEntry.setUser(null);
+            return updatedEntry;
         }
         return null;
     }
 
-    public List<Entry> findAll() {
-        return entryRepository.findAll();
+    public List<Entry> findAll(String username) {
+        return entryRepository.findEntriesByUser_Username(username);
     }
 }
